Reorganize Config Code
