+++ MLBean/projects/conv_pooled/all_config.py
@@ -1,5 +1,6 @@
 from typing import Optional
 import platform
+import pathlib
 
 from MLBean.configs.config_base import BaseConfig, UnionLikeConfig
 from MLBean.data.dataset import DatasetConfig
@@ -36,10 +37,10 @@ def get_basic_all_config() -> AllConfig:
   emb_dims = 256
   transformer_dims = 1024
   num_heads = 16
-  num_layers = 6
+  num_layers = 8
   output_layers = 3
-  group_size = 4
-  conv_kernel_size = 4
+  group_size = 2
+  conv_kernel_size = 6
   return AllConfig(
     model=ConvModelConfig(
       conv_pooled_transformer=ConvPooledTransformerConfig(
@@ -74,16 +75,28 @@ def get_basic_all_config() -> AllConfig:
       ),
     ),
     training=TrainerConfig(
-      num_steps=100000,
+      num_steps=1_000_000,
       log_interval=100,
       eval=EvalConfig(interval=500, steps=100),
       checkpointing=CheckpointingConfig(interval=1000) if platform.system() == "Darwin" else None,
     ),
     optimizer=OptimizerConfig(adamw=AdamWConfig(lr=0.00001)),
     dataset_train=DatasetConfig(
-      batch_size=2,
-      trunc_len=1024,
+      batch_size=4,
+      trunc_len=512,
       group_size=group_size,
       special_tokens_at_end=False,
     ),
   )
+
+
+def get_all_config(chkpt_dir: pathlib.Path) -> AllConfig:
+  all_config = get_basic_all_config()
+  if (chkpt_dir / "all_config.json").exists():
+    all_config = AllConfig.json_load(chkpt_dir / "all_config.json")
+    print(f"succesfully loaded config from {chkpt_dir / 'all_config.json'}")
+    return all_config
+  else:
+    all_config.json_dump(chkpt_dir / "all_config.json")
+    print(f"Saved config to {chkpt_dir / 'all_config.json'}")
+    return all_config
